requirements.txt
requests==2.9.1
python-dateutil==2.4.1
future==0.16.0

nose==1.3.7
coverage==4.2
mock==2.0.0
PyMySQL